use a json file for the babel caching
We were generating python code for the i18n config in the frontend.
Now we serialize to JSON and parse it instead, the python "API" stays
the same.

The problems with generating python code were:
1) differences in the output when running `npm run build` (different
   formatting depending on the local system)
2) different ordering of the dicts depending on the local system
3) rewrites in the autopep8 pre-commit hook

With this commit we rewrite it to indented & ordered JSON so differences
are simpler to review. Also the serialization is using 100% standard
tooling instead of code generation.