[FIX] Re-enable parser caching (#4543)
Since about a week, the page load times in Hedy have gone up by a lot. My best guess is that the cause is disabling the parser caching in https://github.com/hedyorg/hedy/pull/4504.

This re-enables the parser cache to see if this fixes the load times.

**How to test**

Merge, and eyeball the page load times in Heroku.