[REFACTOR] script to reformat the yaml files in content/
We also add the yaml rewriter as a pre-commit hook.

Also remove an empty readme file in the content dir.

Note that there is significant whitespace in the adventures/ directory.
We don't remove leading and trailing whitespace there.

Note on ruamel / yaml 1.2:

We use ruamel for yaml 1.2 compliant parsing

pyyaml does not support yaml 1.2 (yet) and ruamel does not support the
nicer custom yaml.dump format options that pyyaml uses. Not to worry, we
use the best of both worlds ;)

The difference is that in yaml 1.1 No/Yes etc were parsed to boolean
values. We instead interpret it according to yaml 1.2 and thus parse
them as strings like "No"