[UI, FIX] adjusts hide and disables save warning popup in slides (#4556)
Fixes #4513. Also disables the save pop up which is quite annoying in slides mode.

As per https://github.com/hedyorg/hedy/pull/4363 I should not touch the front-end without supervision, so it is probably a good idea if @jpelay checks if this indeed works, and maybe to add a test for the slides and/or for raw? :D