[FIX] order of slides' levels & pages for Turkish only (#4552)
**Description**

I manually reordered slide pages and levels for Turkish language. I first ordered page numbers in slides, then reordered levels. Also added a little touch to the last page of level 2. So from the file's perpective the things are alright.

Please check if editing the file this way is alright from the project's perspective and that of Weblate.

**Fixes** the ordering issue on https://github.com/hedyorg/hedy/issues/4480

- [x] Links to an existing issue or discussion