[CHORE] run snippet generation on changes
Before this change we ran the snippet generation daily. This resulted
in commits even when there was no activity. As the snippets file is a
bit big and has binary contents, the history of the repository grew a
lot.

Now we switch to only updating the hash generation when one of the files
that invalidates the entire cache is updated.

We do not do this on PRs as we can not/ do not want to push commits to
forked repositories.

After we merge a PR to main, and it had one of the files affecting
snippet generating updated, we make a separate commit that updates the
snippet hashes file.