[REFACTOR] speed up unit tests with new cache
This speeds up the unit tests:
- non-cached run: 1.5 hours -> 35 minutes
- cached run: 11 minutes -> 7 minutes

We previously used a snippet hash pickled file to see if snippets had
been executed successfully already on this version of hedy.

This was fine but had two downsides:
- the snippet file changed frequently and was checked into git, leading
  to frequent non-functional changes
- using multiple processes to speed up the CI run was not possible
  as each process would try to overwrite the snippet hash file

Now we introduce pytest-xdist to speed up the test runs by using 2
processes concurrently. We can attempt to go to 4 but the Github runners
have 2 cores and can go out of memory if we go to more concurrency.

We also use a directory based cache instead of pickling a set. We use
concurrency-friendly OS primitives to see if a file exists.

This cache is .gitignore'd so it's not stored in the repository. It is
also stored as a github cache entry. The key uses the hedy language
definition as an input and then hashes it with sha256sum.

This removes the need for the periodically generating of the snippet
hashes, so that workflow is removed.

As the entire hedy language definition was calculated with every test
run, and this was hashed with md5. We change this a bit to use the more
modern and standard but still fast sha1. The sha1 for the entire
language is only used once to calculate a short digest for the language.
That digest is then concatenated to the snippets to form the indivual
test snippets.