[REFACTOR] use functools.cache for parser cache
We used custom cache for the parsers. This was fine but we can use
out-of-the-box tooling just as well.

functools.cache was introduced in python 3.9 and is a nice one-liner.

However, after this, unit tests were failing as the cache key was
dependent on a "hidden" variable, sourcemap.skip_faulty.

This commit also makes that relationship transparent.