[CHORE] run snippet generation on changes (#4573)
**Description**

Before this change we ran the snippet generation daily. This resulted in commits even when there was no activity. As the snippets file is a bit big and has binary contents, the history of the repository grew a lot.

Now we switch to only updating the hash generation when one of the files that invalidates the entire cache is updated.

We do not do this on PRs as we can not/ do not want to push commits to forked repositories.

After we merge a PR to main, and it had one of the files affecting snippet generating updated, we make a separate commit that updates the snippet hashes file.

**Fixes _issue or discussion number_**

Fixes #4546 .

**How to test**

This is a bit sub-optimal, but you can only test this change with quite some effort (creating a fork etc). However, I did some of these tests already and I think the pragmatic way to test is: read the workflow change critically, see if we agree with the proposed flow, merge it and fix any problem as we go along.

After merging we should check this:
- merge a PR that does NOT affect hedy.py or the grammars -> observe that the action does not run
- also merge a PR that does affect hedy.py or the grammars -> observe that a commit is added after all the tests are executed

**Checklist**
Done? Check if you have it all in place using this list: (mark with x if done)

- [x] Contains one of the PR categories in the name
- [x] Describes changes in the format above
- [x] Links to an existing issue or discussion
- [x] Has a "How to test" section