[FIX] reduce production memory usage by caching parsers to disk (#4581)
## Description

Implement a disk based caching system for the parsers. They take about
0.5 seconds to be constructed which leads to delays and high CPU usage.

We used to cache to RAM but now needed multiple GBs of RAM for all the
parsers. The servers don't like this.

With this PR we cache to disk, by pickling the parsers.

NOTE: this PR contains a commit which belongs in #4574 but was needed for this change too. If it gets merged the commit can be dropped from this PR.

## Link to issue

Discussed this in PR #4576 .

## How to test

- See all CI tests succeeding
- Run the entire pytest suite without this fix locally and observe a +/- 1.5GB memory footprint.
- Run with this patch and see that it should only uses about 400MB, while being roughly as fast to execute
- After deployment to production there should be less out of memory issues.

## Checklist
Done? Check if you have it all in place using this list: (mark with x if done)

- [x] Contains one of the PR categories in the name
- [x] Describes changes in the format above
- [x] Links to an existing issue or discussion
- [x] Has a "How to test" section