[FIX] fix flaky cypress tests
When looking at the errors at https://github.com/hedyorg/hedy/actions/runs/6392606011/job/17350202761?pr=4581
I noticed that going to the logged in page sometimes redirects to the
profile page, meaning the session is already logged in.

Cypress should run with test isolation but I'm not 100% sure how it's
configured. The simplest thing I can think of is to clear the cookies +
session + localstorage whenever we call login. That way we are sure that
we have a fresh session.

I built a mini-test locally that tries to login 100x, which failed every
time at some point. With this change it succeeds.