[FIX] fix the github workflow for generating snippet hashes (#4589)
**Description**

After my attempt to make the snippet generation change-based in #4573 there's good news and bad news. The good is that the trigger works, and a PR that had a new `hedy.py` file triggered the snippet generation. The bad news is that the automatic commit action failed due to a permission issue.

https://github.com/hedyorg/hedy/actions/runs/6409135403/job/17399693158

Reading the workflow logic this does not make sense to me and feels like a bug in the git-auto-commit-action but it is what it is :shrug: . As we had no simple way of testing this we decided to merge anyway and solve problems as we go along. This fixes the issue.

**How to test**

- merge this and see if the next change to grammars and/or hedy.py manages to commit the snippet hashes correctly.

**Checklist**

- [x] Contains one of the PR categories in the name
- [x] Describes changes in the format above
- [x] Links to an existing issue or discussion
- [x] Has a "How to test" section

If you're unsure about any of these, don't hesitate to ask. We're here to help!