[DOCS] Readme Update (#4614)
Hello, in this PR I am adding links to the existing issue templates in the project, to speed up the issue generation process for future contributors. Also add the logo and edit the content by removing all the html to keep only markdown code, also, add thanks to the contributors of this project in the final part.

Attached screenshot of the start:
![image](https://github.com/hedyorg/hedy/assets/80181518/ae0b0c2c-fc82-4a8d-9831-5c22ae8fee7a)

Attached screenshot of the end:
![image](https://github.com/hedyorg/hedy/assets/80181518/13624290-e73d-49c8-a83b-4a417c0eeba1)