Implement a dedicated embedding route (#4615)
In this PR we implement a dedicated route for an embedded code editor. This allows external parties to integrate Hedy more easily into their own website. For example as an iframe. A program can be provided in the url as base64 encoded that allows the embedding to include a starting program. Lastly we allow the setting of a parameter to only show the code editor or also show the output and run button.