[FIX] Fix inline keyword language switcher (#4654)
Switching the keyword language produced "converting circular JSON" error message.

The reason was that we were trying to post the entire editor object as JSON to the server, instead of just the code from the editor (bug introduced in the following refactoring #4341).

**How to test**

Log in, set your language to non-English and your preferred keyword language to non-English.

Then use the language switcher on a code page.