Polish translations for some part of adventures (#4657)
I've extended the Polish translations so adventures up to level 3 (including 3) is fully covered.

Sorry for not following the suggested structure of the PR, but couldn't find one for translation work.

I didn't use Weblate because with editing file and running the website with Docker I feel more in control, with Weblate I somehow couldn't map translation to the actual content