[LANGUAGE, UI] Integrate Skulpts debugger to Hedy (#4465)
**Description**

Skulpt has an integrated debugger that is shipped as a separate entity and is envisioned to work like pdb. The basis that is there, however, can help us a great deal in getting our own debugger to work. For this, first we need to integrate the debugger that comes with Skulpt and edit it so it can work with our code. This will not affect Skulpt, since we are just extending it, so any problems that might come with the debugger will not affect the normal execution of code.

However, by doing this, we are taking the responsibility to maintain this piece of code, but I think is very much worth the cost. 

The code in this PR will most likely be a little ugly since this is a proof of concept about getting this to work, and then I'll tackle the task of improving the code and making it adhere to our standards.

**fixes #4427**