[FIX] Fix error messages in debugger, and temporary fix for cypress tests (#4686)
After merging #4465 some Cypress tests were failing. This was because an ad-hoc calcutation I did to figure out the current debugging line. It worked fine locally and even in the tests in the PR, but for some reason failed later on, this PR deabilates that check and I will work on it later this week.

Also, after testing in Alpha I found out I wasn't showing the error messages when an error ocurred during execution. 

How to test?
Go to level 3 and write this code:
```
l is 1, 2
print l at 3
```
Then debug it and when you reach the second line this error should be shown:

![image](https://github.com/hedyorg/hedy/assets/45865185/62836f09-e3e9-403b-bdb0-b99394bd718b)