[FIX] Revert the `filtered_programs_for_user` change
In #4644, the call signature of `filtered_programs_for_user` was
changed, and the call `public_programs_for_user` removed.

Unfortunately, there is no index on the pair `(username, public)`.
Instead, we have indexes on `(username)` and on `(public, date)`.

The first is used to retrieve all programs by a single user, the second
is used to retrieve all public programs (for the "Explore" page). To get
the public programs for a single user, instead what we do is retrieve
all programs for a single user, and filter them down in Python.

Because of a bug, trying to retrieve programs by `{ "username": "henk",
"public": 1 }` uses the `(public, date)` index. However, the database
layerbut does not detect that the field `date` is missing and that the
field `username` is unnecessarily passed. The query incorrectly succeeds
on the local database emulation layer, and when passed to an actual
DynamoDB database the server returns an error.

This currently breaks the profile page, as well as some other pages
probably.

Revert the changes from #4644 that lead to this problem, restoring
the `public_programs_for_user` method and undoing the change to
`filtered_programs_for_user`.