[FIX] Fix the `filtered_programs_for_user` call (#4724)
In #4644, the call signature of `filtered_programs_for_user` was changed to replace the function `public_programs_for_user`. The goal was to be able to supply the query `get_many({ "username": "henk", "public": 1 })`, with the goal to retrieve all public programs by user `henk`.

Unfortunately, there is no index on the pair `(username, public)`. Instead, we have indexes on `(username)` and on `(public, date)`.

The first is used to retrieve all programs by a single user, the second is used to retrieve all public programs (for the "Explore" page). To get the public programs for a single user, instead what we do is retrieve all programs for a single user, and filter them down in Python.

Because of a bug in the database layer, trying to retrieve programs by `{ "username": "henk", "public": 1 }` uses the `(public, date)` index. However, the database layer does not detect that in that particular query the field `date` is missing, and that the field `username` is unnecessarily passed. The query incorrectly succeeds in the local database emulation mode so this goes undetected during testing, and when passed to an actual DynamoDB database the server returns an error.

This currently breaks the profile page, as well as some other pages probably.

Add `public` as an optional keyword argument to `filtered_programs_for_user` and do the filtering locally instead.

**How to test**

The following pages should work:

* `/my-profile`
* `/programs`
* `/class/<class_id>/programs/<username>`