[CHORE] Use Skulpt version from our fork (#4760)
**Description**

On #4465 we added a debugger to Hedy that needed to use some changes on the Skulpt library. For that I added a Skulpt version that I took from my own fork; now that we have our fork of Skulpt https://github.com/hedyorg/skulpt I'll use this oportunity to delete some `console.log`s that I left there and also explicitly leave a trace to point to a single PR that changed the version of Skulpt we are using.