[CHORE] Unwrap Babel translations files (#4767)
**Description**

A while ago we opened #4501 to add an add-on to Weblate so it wouldn't wrap files, we also changed the documentation to specify a `--no-wrap` option when extracting the labels. However, we are still running into problems where it seems like the files are actually wrapped. 

Looking into the git blame for the files, it seems like some lines haven't been changed in months, so maybe we actually need to change them manually by unwrapping all of them, and then send them to Weblate and see what happens?