[FIX] Improve reliability of second teacher Cypress test
I tried improving the reliability of the cypress tests for second
teachers, because I selfishly want my own PR merged and it blocked on
this 😈.

I made the following changes:

The tests used to create a new class, but then always clicked into the
*first* class they could find on the teachers page instead of the class
they created. Depending on what names get generated, this could be a
class left by a previous test and there's no knowing what state that
class is in. Use `navigateToClass(className)`.

Replace `should()` with `contains()` in most places. `contains()` is
a query, meaning it *looks and waits* for elements containing that string. On
the other hand, `should()` is an assertion: on the elements your
query is already focused on, assert that they have this text. This
prevents accidentally matching the object saying `teacher2` (due to
timing) while the object containing `teacher3` will appear right next
to it a few milliseconds later, but your query had already settled
on a certain set of DOM elements and it will have moved on already.

Replace some JavaScript navigating the invite table with a selector
involving `.parent()`: we find the row containing the teacher we're
interested in, then use `.parent().children()` to find other
elements *in the same row*.