[FIX] Improve reliability of second teacher Cypress test (#4805)
I tried improving the reliability of the cypress tests for second teachers, because I selfishly want my own PR merged and it blocked on this 😈.

I made the following changes:

### Class selection

The tests used to create a new class, but then always clicked into the *first* class they could find on the teachers page instead of the class they created. Depending on what names get generated, this was not necessarily the fresh class we just generated, but could be any class left by a previous test and there's no knowing what state that class is in. Use `navigateToClass(className)`.

### Should vs contains

Replace `should()` with `contains()` in most places. `contains()` is a query, meaning it *looks and waits* for elements containing that string. On the other hand, `should()` is an assertion: on the elements your query is already focused on, assert that they have this text. The difference comes up when timing comes into the picture. Here's an example:

Let's say we have an element with text `teacher2`, and after we click a button `teacher3` is going to be added to the list. However, this may take a few milliseconds.

```ts
cy.get('.teacher_link')                   // <--- (Q)
  .should('include.text', 'teacher3');    // <--- (A)
```

In this example, the query `Q` is "find me all elements with class `teacher_link`". If this executes *after* `teacher3` has been added to the list, then that query will match two elements, and the assertion `A` can see that, indeed, one of them has the text `teacher3`.

If, on the other hand, this executes *before* `teacher3` has been added to the list, then the query is still satisified. It can still find at least one element with the class `teacher_link`, and it will move on to the assertion. Even if the `teacher3` element appears a millisecond later, it's too late: the query has stopped waiting.

If we write the cypress statement like this:

```ts
cy.get('.teacher_link')     
   .contains('teacher3');     
```

Then now *both* statements are part of the query! Cypress will look and wait for an element that is BOTH a `.teacher_link` as well as has text `teacher3`. There isn't one, until there is. What won't happen, is that Cypress will settle on the wrong set of elements too early.

### DOM navigation

Replace some JavaScript navigating the invite table with a selector involving `.parent()`: we find the row containing the teacher we're interested in, then use `.parent().find()` to find other elements *in the same row*.

To finish the assertion, I've needed to move the test up, so that we revoke teacher3's invitation before we accept teacher2's invitation (otherwise at the end the table is altogether gone and the `cy.get()` fails).

**How to test**

GitHub will test for us!