[BUG FIX] Fix embedded editor with CodeMirror (#4821)
**Description**
In this PR we fix some issues that were the result of the migration from ace to CodeMirror. The embedded editor view was not updated yet and still relied on the ace editor. This resulted in some rendering and postMessage issues, making it unusable when embedding.

**Fixes _issue or discussion number_**
This PR fixes #4820

**How to test**
Navigate to the following link when running locally, all should work as expected:

http://127.0.0.1:8080/1/?lang=nl&run=true&readOnly=false&program=cHJpbnQgaGFsbG8gd2VyZWxkIQ==