[UI, FIX] Fixes placing of checkbox and eye icon on grid (#4834)
**Description**

Changes the table elements to be each a table layout of 3 columns, where the first column is empty, for alignment, the second one is the checkbox and the third one is the eye. This fixes the overflowing behavior that the elements of this table presented.

Worth noting that this page will be moved from here to the dashboard, but we need this fix shipped because it's very annoying!

**Fixes [this comment](https://github.com/hedyorg/hedy/issues/4580#issuecomment-1803884643) on #4580**

**How to test**

* Make sure that some students have adventures done for a particular class.
* Go to the class and click on Overview of programs per adventure.
* Zoom-in and out and see that the elements do not overflow or mess with each other