[FIX] Proper marking of flat compound statements while debugging (#4766)
**Description**

Changes the regex that was used to group the different parts of the flat compound statements to be able to properly color nested flat statements. Also changes the way the line is computed.

A nested flat statement are `if` with `repeat` inside or viceversa. You cannot place an `if` inside a `if`, but you can place a 
`repeat` inside a `repeat`.

Even though we say these statements are flat, it's possible to add newlines between them. And this is where the problem lied. When these statements where split between lines, the code did not account for beginning and end of statements properly.


**Fixes #4725 **

**How to test**

I added a couple of tests!