[FIX, CHORE] Updates @codemirror/view (#4803)
**Description**

Updates the version of @codemirror/view, which is the component of CodeMirror that handles the UI of the library, to the last version. The issue we were experiencing with the gutter was actually caused by a bug in the library.

**Fixes #4776**

**How to test**

Go to level 12 and delete all of the code in the editor, then copy and paste this code:

```
define simple_function_1 with parameter
    print 'simple_function_1 - 1'
    m = 'simple_function_1 - 2'
    print m
define simple_function_2 with param
    print 'simple_function_2 - 1'
    print param
define simple_function_3 with param_a, param_b, param_c
    if param_a = 'A'
        print 'simple_function_3 - 1'
        print param_b
    else
        print 'simple_function_3 - 2'
        if param_a = 'B'
            print 'simple_function_3 - 2A'
            print param_b
        else
            print 'simple_function_3 - 2B'
            print param_c
a = 'test1'
print call simple_function_3 with 'A', a, 1.0
call simple_function_3 with 'B', a, 1.0
call simple_function_3 with 'C', a, 1.0
```

On this branch version the gutter should be shown correctly, as opposed to alpha.