[FEAT] Check field types in emulation layer
There is currently a problem with a table that has an index on a
certain field, but is trying to insert values of a type not matching
that index. It fails in Dynamo, which doesn't allow that.

However, the local emulation layer doesn't catch that, because it
doesn't know anything about types that need to be checked.

Add a required `types` argument to `Table`, which must declare at least
the key fields (of the table itself and of the indexes), and can
optionally declare the types of other fields. It will be an error to
insert values of the wrong type into the database.

This may look like an easy feature, but it has some edge cases you
need to be aware of:

* First of all, it relies on users declaring the indexes in code; they
  can choose NOT to declare the indexes in code, and there may still
  be an error after deploying.
* Second of all: just because you declare a type on a fields saying what
  may be inserted into a table going forward, this can not make any
  guarantees on data already in the table. Don't use this feature
  to make any assumptions on data you read out of the table, only
  use it to sanity-check new data going into it!

We may decide this change is not worth it... (for example, it would
not have caught the current problem, as the index was there in Dynamo
but not declared in code, so this feature wouldn't have caught it).