[CHORE] move exceptions into isvalid (#4838)
First step towards #4836 

This PR cleans up the handling of exceptions a lot! As I explained in #4836 I think I initially was planning to create a function that would return a boolean, is this tree valid or not, with a transformer. But as the code got more advanced, and more experience people worked in it, we started to use exceptions more and it became a bit of a mess.

This trigger for this specific refactoring was there was not an easy way to call one function to detect if a tree was valid, which is needed for the new translation mechanism. So it does not fully close #4836 as I still have to call it on the translation part but this is already quite big and hard to review so let's do that separately.

**How to test**
Verify nothing breaks :)