[UI] Add information tooltip for slides (#4870)
**Description**

Adds a button for displaying information about the slides in the For Teachers page. I didn't display it in a tool-tip because of advice from this page: https://www.smashingmagazine.com/2021/02/designing-tooltips-mobile-user-interfaces/ Since the information was too long I preferred to add it as a separate section where it doesn't intrude with the other elements.


**Fixes #4750**


**How to test**

Go to http://0.0.0.0:8080/for-teachers and click the information button next to slides :)