[FIX] Remove fixed height for editor
We used to set fixed pixel heights for the (CodeMirror) editor; one
of the things we would do is measure the height of the container
during editor initialization, and then force the editor to have that
height.

On the slides page, where we would load a lot of pages in `<iframe>`s
that would start out hidden, the calculated height would sometimes
be measured as `0px`, which would then be copied onto the editor.
The end result would be that the editor would appear empty on some
pages.

Instead, in this PR: remove all height calculations. We will set a
height on the entire editor grid (containing the editor and output box,
and navigation buttons), and add the CSS declarations necessary to make
the editor automatically resize to take up available remaining
space, after the fixed elements are taken away (like the buttons
at the bottom, the error and success popups, the "handed in" warning
box, etc).

This makes it no longer necessary to recalculate the height of the
editor every time one of those popups is shown or hidden.